+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.2 (UNRELEASED)
 ### Changed
 - Removed dependency on `tomli` to simplify installation.
+- Improve error messages on invalid inputs to `rust_extensions` keyword. [#203](https://github.com/PyO3/setuptools-rust/pull/203)
 
 ## 1.1.1 (2021-12-01)
 ### Fixed